+++ json_serializable/test/supported_types/input.type_bigint.g.dart
@@ -10,20 +10,28 @@ part of 'input.type_bigint.dart';
 
 SimpleClass _$SimpleClassFromJson(Map<String, dynamic> json) => SimpleClass(
       BigInt.parse(json['value'] as String),
+      json['withDefault'] == null
+          ? _defaultValueFunc()
+          : BigInt.parse(json['withDefault'] as String),
     );
 
 Map<String, dynamic> _$SimpleClassToJson(SimpleClass instance) =>
     <String, dynamic>{
       'value': instance.value.toString(),
+      'withDefault': instance.withDefault.toString(),
     };
 
 SimpleClassNullable _$SimpleClassNullableFromJson(Map<String, dynamic> json) =>
     SimpleClassNullable(
       json['value'] == null ? null : BigInt.parse(json['value'] as String),
+      json['withDefault'] == null
+          ? _defaultValueFunc()
+          : BigInt.parse(json['withDefault'] as String),
     );
 
 Map<String, dynamic> _$SimpleClassNullableToJson(
         SimpleClassNullable instance) =>
     <String, dynamic>{
       'value': instance.value?.toString(),
+      'withDefault': instance.withDefault?.toString(),
     };